Ask for creation of uid/gid if they don't exist

+++ setup-phase-1-2.sh
@@ -11,6 +11,19 @@ function phase1_lilipod_distrobox_install() {
       exit 1
    }
 
+   if [ -s "/etc/subgid" ] || [ -s "/etc/subuid" ]; then # only manjaro had those files missing so far
+      echor "You seem to be missing subgid or subuid which is required for rootless containers to work"
+      echog "I can create those for you, next sudo prompt will ask for permissions from you to do so."
+      echog "Confirm if you're okay with proceesing (y/N)"
+      read -r confirm
+      if [[ $confirm != "y" ]] && [[ $confirm != "Y" ]]; then
+         echor "User has not confirmed creation of gid/uid, can't complete installation, exiting."
+         exit 1
+      fi
+      echo "$USER:100000:65536" | sudo tee /etc/subuid
+      echo "$USER:100000:65536" | sudo tee /etc/subgid
+   fi
+
    echog "Using shipped getsubids"
    cp "../getsubids" "$prefix/bin/"
    if [[ "$(which getsubids)" != "$prefix/bin/getsubids" ]]; then
